apps/jsonrpc: avoid compiler warning by casting to signed

+++ apps/jsonrpc/RpcPeer.cpp
@@ -238,11 +238,11 @@ int JsonrpcNetstringsConnection::netstringsRead() {
 	  return REMOVE;
 	}
 
-	if ((rcvd_size<0 && errno == EAGAIN) || 
-	    (rcvd_size<0 && errno == EWOULDBLOCK))
+	if (((ssize_t)rcvd_size<0 && errno == EAGAIN) || 
+	    ((ssize_t)rcvd_size<0 && errno == EWOULDBLOCK))
 	  return CONTINUE; // necessary?
 
-	if (rcvd_size<0) {
+	if ((ssize_t)rcvd_size<0) {
 	  INFO("socket error on connection [%p/%d]: %s\n",
 	       this, fd, strerror(errno));
 	  close();
@@ -296,7 +296,7 @@ int JsonrpcNetstringsConnection::netstringsRead() {
 }
 
 int JsonrpcNetstringsConnection::netstringsBlockingWrite() {
-  if (msg_size<0) {
+  if ((ssize_t)msg_size<0) {
     close();
     return REMOVE;
   }
@@ -327,12 +327,12 @@ int JsonrpcNetstringsConnection::netstringsBlockingWrite() {
 			  0
 #endif
 			  );
-    if ((written<0 && (errno==EAGAIN || errno==EWOULDBLOCK)) ||
+    if (((ssize_t)written<0 && (errno==EAGAIN || errno==EWOULDBLOCK)) ||
 	written==0) {
 	usleep(SEND_SLEEP);
 	continue;
     }
-    if (written<0) {
+    if ((ssize_t)written<0) {
       if (errno == ECONNRESET) {
 	DBG("closing connection [%p/%d] on peer hangup\n", this, fd);
 	close();
